Started the SDL_Texture rendering implementation and improved some features - Converted the rotation to a float - Created the SDL_Renderer for the texture rendering - Splitted the RenderObject function - Created the SurfaceImage and the TextureImage

+++ PathfindingAlgorithmsProject/PathfindingAlgorithmsProject/GameObject.h
@@ -121,7 +121,7 @@ public:
 	/*
 	Return the GambeObject's rotation (counting all the parents' rotations)
 	*/
-	short GlobalRotation() const;
+	float GlobalRotation() const;
 	/*
 	Return the GambeObject's alpha (counting all the parents' alphas)
 	*/
